DocumentTime: Store stuff as 32bit ints instead of 64
Saves space since we really do not need to store such large numbers.

+++ src/engine/document.h
@@ -69,9 +69,9 @@ public:
     void setContentIndexing(bool val);
     bool contentIndexing() const;
 
-    void setMTime(quint64 val) { m_mTime = val; }
-    void setCTime(quint64 val) { m_cTime = val; }
-    void setJulianDay(quint64 val) { m_julianDay = val; }
+    void setMTime(quint32 val) { m_mTime = val; }
+    void setCTime(quint32 val) { m_cTime = val; }
+    void setJulianDay(quint32 val) { m_julianDay = val; }
 
     void setData(const QByteArray& data);
 
@@ -91,9 +91,9 @@ private:
     QByteArray m_url;
     bool m_contentIndexing;
 
-    quint64 m_mTime;
-    quint64 m_cTime;
-    quint64 m_julianDay;
+    quint32 m_mTime;
+    quint32 m_cTime;
+    quint32 m_julianDay;
     QByteArray m_data;
 
     friend class Database;
